feat: 컴포넌트 리스트 레이어 추가 후 스택타입리스트 이동 및 스타일 일부 개선
