+++ src/component_list/StackTypeList/index.module.scss
@@ -1,16 +1,19 @@
 .stack_type {
   display: flex;
   gap: 10px;
-  margin-bottom: 10px;
 }
 
 .stack_type_btn {
   font-size: x-large;
   font-weight: 600;
-  border-radius: 41px;
+  border-radius: 10px;
   padding: 5px 20px;
 
-  &:active {
-    background-color: #737373;
+  &.selected {
+    background-color: #bcbcbc;
   }
+}
+
+.info {
+  font-size: small;
 }
